producer/requirements.txt
praw==7.7.1
kafka-python==2.0.2
python-dotenv==1.0.0